cleanup and separate concerns in media module
